python/narrative_stack/pyproject.toml
[project]
name = "narrative-stack-local"
version = "0.1.0"
description = "Python utilities for Narrative Stack"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py==2.2.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.15",
    "aiosignal==1.3.2",
    "alembic==1.15.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "appnope==0.1.4",
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==21.2.0",
    "arrow==1.3.0",
    "asttokens==3.0.0",
    "async-lru==2.0.5",
    "attrs==25.3.0",
    "babel==2.17.0",
    "beautifulsoup4==4.13.3",
    "bleach==6.2.0",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "colorlog==6.9.0",
    "comm==0.2.2",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "datasets==3.4.1",
    "debugpy==1.8.13",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "dill==0.3.8",
    "dotenv==0.9.9",
    "et_xmlfile==2.0.0",
    "executing==2.2.0",
    "fastjsonschema==2.21.1",
    "filelock==3.13.1",
    "fonttools==4.57.0",
    "fqdn==1.5.1",
    "frozenlist==1.5.0",
    "fsspec==2024.6.1",
    "grpcio==1.71.0",
    "h11==0.14.0",
    "hdbscan==0.8.40",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "huggingface-hub==0.29.3",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==9.0.2",
    "ipython_pygments_lexers==1.1.1",
    "ipywidgets==8.1.5",
    "isoduration==20.11.0",
    "jedi==0.19.2",
    "Jinja2==3.1.4",
    "joblib==1.4.2",
    "json5==0.10.0",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "jupyter-console==6.6.3",
    "jupyter-events==0.12.0",
    "jupyter-lsp==2.2.5",
    "jupyter_client==8.6.3",
    "jupyter_core==5.7.2",
    "jupyter_server==2.15.0",
    "jupyter_server_terminals==0.5.3",
    "jupyterlab==4.3.6",
    "jupyterlab_pygments==0.3.0",
    "jupyterlab_server==2.27.3",
    "jupyterlab_widgets==3.0.13",
    "kiwisolver==1.4.8",
    "lightning-utilities==0.14.2",
    "llvmlite==0.44.0",
    "Mako==1.3.9",
    "Markdown==3.7",
    "MarkupSafe==2.1.5",
    "matplotlib==3.10.1",
    "matplotlib-inline==0.1.7",
    "mistune==3.1.3",
    "mpmath==1.3.0",
    "multidict==6.2.0",
    "multiprocess==0.70.16",
    "mysql-connector-python==9.2.0",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "networkx==3.3",
    "notebook==7.3.3",
    "notebook_shim==0.2.4",
    "numba==0.61.2",
    "numpy==2.1.2",
    "openpyxl==3.1.5",
    "optuna==4.2.1",
    "overrides==7.7.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.0.0",
    "platformdirs==4.3.7",
    "prometheus_client==0.21.1",
    "prompt_toolkit==3.0.50",
    "propcache==0.3.0",
    "protobuf==6.30.1",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure_eval==0.2.3",
    "pyarrow==19.0.1",
    "pycparser==2.22",
    "pydantic==2.11.4",
    "pydantic_core==2.33.2",
    "Pygments==2.19.1",
    "pynndescent==0.5.13",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-json-logger==3.3.0",
    "pytorch-lightning==2.5.1",
    "pytz==2025.1",
    "PyYAML==6.0.2",
    "pyzmq==26.3.0",
    "RapidFuzz==3.12.2",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rpds-py==0.23.1",
    "safetensors==0.5.3",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "Send2Trash==1.8.3",
    "sentence-transformers==3.4.1",
    "setuptools==70.2.0",
    "simd-r-drive-ws-client>=0.10.0a0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "SQLAlchemy==2.0.39",
    "stack-data==0.6.3",
    "sympy==1.13.1",
    "tensorboard==2.19.0",
    "tensorboard-data-server==0.7.2",
    "terminado==0.18.1",
    "threadpoolctl==3.6.0",
    "tinycss2==1.4.0",
    "tokenizers==0.21.1",
    "torch==2.6.0",
    "torchaudio==2.6.0",
    "torchmetrics==1.7.0",
    "torchvision==0.21.0",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.50.0",
    "typer>=0.16.0",
    "types-python-dateutil==2.9.0.20241206",
    "typing-inspection==0.4.0",
    "typing_extensions==4.12.2",
    "tzdata==2025.1",
    "umap-learn==0.5.7",
    "uri-template==1.3.0",
    "urllib3==2.3.0",
    "wcwidth==0.2.13",
    "webcolors==24.11.1",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "Werkzeug==3.1.3",
    "widgetsnbextension==4.0.13",
    "xxhash==3.5.0",
    "yarl==1.18.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
exclude = [
    "notebooks"
]
# TODO: Add stricter lint rules: https://docs.astral.sh/ruff/rules/
# select = ["E", "F", "G", "B"]
# extend-select = ["C", "I", "N", "UP"] 

